Refacto study listing view

+++ antarest/storage/business/raw_study_service.py
@@ -260,7 +260,12 @@ class RawStudyService:
         metadata.path = str(path_study)
         return metadata
 
-    def copy_study(self, src_meta: RawStudy, dest_meta: RawStudy) -> RawStudy:
+    def copy_study(
+        self,
+        src_meta: RawStudy,
+        dest_meta: RawStudy,
+        with_outputs: bool = False,
+    ) -> RawStudy:
         """
         Copy study to a new destination
         Args:
@@ -277,7 +282,7 @@ class RawStudyService:
         shutil.copytree(src_path, dest_path)
 
         output = dest_path / "output"
-        if output.exists():
+        if not with_outputs and output.exists():
             shutil.rmtree(output)
 
         _, study = self.study_factory.create_from_fs(
